dermacare-backend/requirements.txt
langchain-iris
testcontainers-iris
llama-iris
llama-index-legacy
sentence-transformers
langchain
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel
setuptools
kagglehub
langchain-experimental
pillow
open_clip_torch
torch
matplotlib
jupyter